Discontinue printing message when UID != root
We want to discourage running the program as root and instead encourage
creating a user with access to input devices instead of the whole
system.

However, a user with access to the input group can write a
keylogger (see eventcap), which is why we should consider telling users
to avoid giving any software extremely high privileges, including this
one.